Rename Character package all lowercase for review #677.
